# core/config/openapi.yml
...
openapi: 3.1.0
...
components:
  schemas:
    AppUserModel:
      type: object
      required:
      - id
      - firstName
      - lastName
      - email
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
...
    LocalDate:
      type: string
      format: date
      examples:
      - 2022-03-10
...
    OffsetDateTime:
      type: string
      format: date-time
      examples:
      - 2022-03-10T12:15:50-04:00
    OrderModel:
      type: object
      required:
      - id
      - fromDate
      - toData
      - hotelId
      - userId
      - roomTypeId
      - createdAt
      - updatedAt
      - removed
      properties:
        id:
          type: integer
          format: int64
        fromDate:
          $ref: "#/components/schemas/LocalDate"
        toData:
          $ref: "#/components/schemas/LocalDate"
        hotelId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        roomTypeId:
          type: integer
          format: int64
        createdAt:
          $ref: "#/components/schemas/OffsetDateTime"
        updatedAt:
          $ref: "#/components/schemas/OffsetDateTime"
        removed:
          type: boolean
...
    SignInRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
        password:
          type: string
    SignUpRequest:
      type: object
      required:
      - firstName
      - lastName
      - email
      - password
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
    UserModel:
      type: object
      required:
      - id
      - firstName
      - lastName
      - email
      - role
      - createdAt
      - updatedAt
      - removed
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        createdAt:
          $ref: "#/components/schemas/OffsetDateTime"
        updatedAt:
          $ref: "#/components/schemas/OffsetDateTime"
        removed:
          type: boolean
    UserRole:
      type: string
      enum:
      - admin
      - user
    UserSaveRequest:
      type: object
      required:
      - firstName
      - lastName
      - email
      - password
      - role
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
          pattern: \S
        lastName:
          type: string
          pattern: \S
        email:
          type: string
        password:
          type: string
          pattern: \S
          default: qwerty
        role:
          type: string
          pattern: \S
          default: user
    UserSaveResponse:
      type: object
      required:
      - id
      properties:
        id:
          type: integer
          format: int64
    UserSession:
      type: object
      required:
      - token
      - created
      - expires
      - id
      - firstName
      - lastName
      - email
      - role
      properties:
        token:
          type: string
        created:
          $ref: "#/components/schemas/OffsetDateTime"
        expires:
          $ref: "#/components/schemas/OffsetDateTime"
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
  securitySchemes:
    SecurityScheme:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication
paths:
...
  /api/auth/sign-in:
    post:
      operationId: signIn
      tags:
      - Auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInRequest"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Sign In
  /api/auth/sign-up:
    post:
      operationId: signUp
      tags:
      - Auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpRequest"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Sign Up
  /api/core/app/users/{id}:
    get:
      operationId: getUserById
      tags:
      - User
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUserModel"
      summary: Get User By Id
...
  /api/core/orders/user/{userId}:
    get:
      operationId: getOrdersByUserId
      tags:
      - Order
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderModel"
      summary: Get Orders By User Id
...
  /api/core/users:
    post:
      operationId: saveUser
      tags:
      - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSaveRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSaveResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Save User
      security:
      - SecurityScheme:
        - admin
    get:
      operationId: getAllUsers
      tags:
      - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserModel"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get All Users
      security:
      - SecurityScheme:
        - admin
  /api/core/users/me:
    get:
      operationId: getMe
      tags:
      - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSession"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get Me
      security:
      - SecurityScheme: []
  /api/core/users/{id}:
    get:
      operationId: getUserById
      tags:
      - User
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserModel"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get User By Id
      security:
      - SecurityScheme:
        - admin
    delete:
      operationId: deleteUserById
      tags:
      - User
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Delete User By Id
      security:
      - SecurityScheme:
        - admin
info:
  title: main-app API
  version: 1.0.0-SNAPSHOT